[feat] (IMPORTANT UPDATE) function prototype (5) - chatroom save information - (Filter and Save to DB Controller) friends calculation (bug: repeat pairs & order) - dev with old DB schema

+++ demo/src/main/java/com/example/demo/model/user/User.java
@@ -12,7 +12,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String accountName;
     private String nickname;
@@ -25,7 +25,7 @@ public class User {
         // no-argument constructor
     }
 
-    public User(Long id, String accountName, String nickname, Integer genderId) {
+    public User(Integer id, String accountName, String nickname, Integer genderId) {
         this.id = id;
         this.accountName = accountName;
         this.nickname = nickname;
